Full room and role maintainer functionality
The switch of USE_THREADS is fully used.
Text channel recycling functionality will happen if USE_THREADS is false in the config. Members will receive a role for their room when the rooms are made. The role maintainer removes roles that members of the server should no longer have because their old mogi lifetime has ended.